products and categories are working for the most part

+++ capstone-starter/src/main/java/org/yearup/controllers/CategoriesController.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.yearup.data.CategoryDao;
 import org.yearup.data.ProductDao;
 import org.yearup.models.Category;
@@ -26,6 +27,7 @@ public class CategoriesController
 
 
     // create an Autowired controller to inject the categoryDao and ProductDao
+
     @Autowired
     public CategoriesController(CategoryDao categoryDao, ProductDao productDao){
         this.categoryDao = categoryDao;
@@ -33,6 +35,7 @@ public class CategoriesController
     }
 
     // add the appropriate annotation for a get action
+
     @GetMapping("")
     @PreAuthorize("permitAll()")
     public List<Category> getAll()
@@ -42,6 +45,7 @@ public class CategoriesController
     }
 
     // add the appropriate annotation for a get action
+
     @GetMapping("{id}")
     @PreAuthorize("permitAll()")
     public Category getById(@PathVariable int id)
@@ -52,14 +56,18 @@ public class CategoriesController
 
     // the url to return all products in category 1 would look like this
     // https://localhost:8080/categories/1/products
+
     @GetMapping("{categoryId}/products")
     @PreAuthorize("permitAll()")
     public List<Product> getProductsById(@PathVariable int categoryId)
     {
         // get a list of product by categoryId
         try {
-            return productDao.listByCategoryId(categoryId);
-        // return (List<Product>) categoryDao.getById(categoryId);
+            if (productDao.listByCategoryId(categoryId) == null)
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            else {
+                return productDao.listByCategoryId(categoryId);
+            }
 
         } catch (Exception e) {
             throw new RuntimeException(e);
@@ -69,17 +77,20 @@ public class CategoriesController
 
     // add annotation to call this method for a POST action
     // add annotation to ensure that only an ADMIN can call this function
+
     @PostMapping("")
     @ResponseStatus(HttpStatus.CREATED)
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public Category addCategory(@RequestBody Category category)
     {
         // insert the category
-        return categoryDao.create(category);
+        Category newCategory = categoryDao.create(category);
+        return newCategory;
     }
 
     // add annotation to call this method for a PUT (update) action - the url path must include the categoryId
     // add annotation to ensure that only an ADMIN can call this function
+
     @PostMapping("{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @PreAuthorize("hasRole('ROLE_ADMIN')")
@@ -91,14 +102,20 @@ public class CategoriesController
 
     // add annotation to call this method for a DELETE action - the url path must include the categoryId
     // add annotation to ensure that only an ADMIN can call this function
+
     @DeleteMapping("{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public void deleteCategory(@PathVariable int id)
     {
-        // delete the category by id
-        try {
+        try
+        {
+            var category = categoryDao.getById(id);
+
+            if(category == null)
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
             categoryDao.delete(id);
+
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
